// frontend/src/client/types.gen.ts
export type ReserveVisitRequest = {
    patientId?: string;
    doctorId?: string;
    type?: 'GENERAL' | 'SPECIALIST' | 'FOLLOW_UP';
    slotIds?: Array<string>;
};
// ...
export type StaffDto = {
    id?: string;
    role?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    active?: boolean;
    specialization?: string;
    pwz?: string;
    department?: string;
    position?: string;
    workSchedule?: string;
};

export type ScheduleDto = {
    scheduleId?: string;
    doctorId?: string;
    slots?: Array<SlotDto>;
};

export type SlotDto = {
    slotId?: string;
    timeRange?: TimeRange;
    office?: string;
    status?: string;
};

export type AvailableSlotDto = {
    slotId?: string;
    doctorId?: string;
    doctorFirstName?: string;
    doctorLastName?: string;
    specialization?: string;
    startTime?: string;
    endTime?: string;
    office?: string;
};